+++ src/components/types.ts
@@ -1,6 +1,11 @@
 export type CellType = 'white' | 'black' | 'start' | 'goal' | 'obstacle';
 
-export type Panel = {
+export interface Panel {
   id: string;
   cells: CellType[][];
-};
+}
+
+export interface CellTypeConfig {
+  label: string;
+  color: string;
+}
